+++ core/api/core.go
@@ -66,7 +66,7 @@ type coreService struct {
 
 	chainID                  string
 	genesisTime              time.Time
-	hasGenesisTimeAndChainID uint32
+	hasGenesisTimeAndChainID atomic.Bool
 	mu                       sync.Mutex
 
 	netInfo   *tmctypes.ResultNetInfo
@@ -102,7 +102,7 @@ func (s *coreService) LastBlockHeight(
 ) (*protoapi.LastBlockHeightResponse, error) {
 	defer metrics.StartAPIRequestAndTimeGRPC("LastBlockHeight")()
 
-	if atomic.LoadUint32(&s.hasGenesisTimeAndChainID) == 0 {
+	if !s.hasGenesisTimeAndChainID.Load() {
 		if err := s.getGenesisTimeAndChainID(ctx); err != nil {
 			return nil, fmt.Errorf("failed to intialise chainID: %w", err)
 		}
@@ -369,7 +369,7 @@ func (s *coreService) getTendermintStats(
 		return 0, 0, nil, "", apiError(codes.Internal, ErrBlockchainBacklogLength, err)
 	}
 
-	if atomic.LoadUint32(&s.hasGenesisTimeAndChainID) == 0 {
+	if !s.hasGenesisTimeAndChainID.Load() {
 		if err = s.getGenesisTimeAndChainID(ctx); err != nil {
 			return 0, 0, nil, "", err
 		}
@@ -437,7 +437,7 @@ func (s *coreService) getGenesisTimeAndChainID(ctx context.Context) error {
 		return apiError(codes.Internal, ErrBlockchainChainID, err)
 	}
 
-	atomic.StoreUint32(&s.hasGenesisTimeAndChainID, 1)
+	s.hasGenesisTimeAndChainID.Store(true)
 	return nil
 }